api-seg/requirements.txt
numpy==1.16.4
fastapi==0.83.0
uvicorn==0.16.0
folium==0.9.1
geojson==2.4.1
h5py==2.9.0
imageio==2.5.0
imgaug==0.2.6
IPython[all]
Keras==2.2.4
matplotlib==3.1.0
opencv-python==4.1.0.25
pandas==0.23.4
Pillow==5.4.1
python-dotenv==0.20.0
scikit-image==0.15.0
scikit-learn==0.21.2
rasterio==1.2.10
tensorflow==1.13.1